# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Neuralix Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /guilds/{guildId}/welcome:
    get:
      operationId: getWelcomeConfig
      tags: [welcome]
      summary: Get welcome configuration
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Welcome config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WelcomeConfig"
    put:
      operationId: updateWelcomeConfig
      tags: [welcome]
      summary: Update welcome configuration
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WelcomeConfigInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WelcomeConfig"
  /guilds/{guildId}/welcome/test:
    post:
      operationId: testWelcome
      tags: [welcome]
      summary: Send test welcome message
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Test sent

  # GOODBYE
  /guilds/{guildId}/goodbye:
    get:
      operationId: getGoodbyeConfig
      tags: [goodbye]
      summary: Get goodbye configuration
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Goodbye config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoodbyeConfig"
    put:
      operationId: updateGoodbyeConfig
      tags: [goodbye]
      summary: Update goodbye configuration
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoodbyeConfigInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoodbyeConfig"
  /guilds/{guildId}/goodbye/test:
    post:
      operationId: testGoodbye
      tags: [goodbye]
      summary: Send test goodbye message
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Test sent

  # VERIFICATION
  /guilds/{guildId}/verification:
    get:
      operationId: getVerificationConfig
      tags: [verification]
      summary: Get verification config
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verification config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationConfig"
    put:
      operationId: updateVerificationConfig
      tags: [verification]
      summary: Update verification config
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationConfigInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationConfig"
  /verify/{guildId}:
    post:
      operationId: verifyUser
      tags: [verification]
      summary: Verify user for a guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResult"

  # TICKETS
...
components:
  schemas:
...
    WelcomeConfig:
      type: object
      properties:
        id:
          type: integer
        guildId:
          type: string
        enabled:
          type: boolean
        channelId:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        embedEnabled:
          type: boolean
        embedColor:
          type: ["string", "null"]
        embedTitle:
          type: ["string", "null"]
        embedDescription:
          type: ["string", "null"]
        imageEnabled:
          type: boolean
        autoRoleIds:
          type: array
          items:
            type: string
        dmEnabled:
          type: boolean
        dmMessage:
          type: ["string", "null"]
      required: [id, guildId, enabled, embedEnabled, imageEnabled, autoRoleIds, dmEnabled]
    WelcomeConfigInput:
      type: object
      properties:
        enabled:
          type: boolean
        channelId:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        embedEnabled:
          type: boolean
        embedColor:
          type: ["string", "null"]
        embedTitle:
          type: ["string", "null"]
        embedDescription:
          type: ["string", "null"]
        imageEnabled:
          type: boolean
        autoRoleIds:
          type: array
          items:
            type: string
        dmEnabled:
          type: boolean
        dmMessage:
          type: ["string", "null"]
    GoodbyeConfig:
      type: object
      properties:
        id:
          type: integer
        guildId:
          type: string
        enabled:
          type: boolean
        channelId:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        embedEnabled:
          type: boolean
        embedColor:
          type: ["string", "null"]
        embedTitle:
          type: ["string", "null"]
        embedDescription:
          type: ["string", "null"]
        imageEnabled:
          type: boolean
      required: [id, guildId, enabled, embedEnabled, imageEnabled]
    GoodbyeConfigInput:
      type: object
      properties:
        enabled:
          type: boolean
        channelId:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        embedEnabled:
          type: boolean
        embedColor:
          type: ["string", "null"]
        embedTitle:
          type: ["string", "null"]
        embedDescription:
          type: ["string", "null"]
        imageEnabled:
          type: boolean
    VerificationConfig:
      type: object
      properties:
        id:
          type: integer
        guildId:
          type: string
        enabled:
          type: boolean
        roleId:
          type: ["string", "null"]
        minAccountAge:
          type: integer
        antiVpn:
          type: boolean
        antiAlt:
          type: boolean
        antiBot:
          type: boolean
      required: [id, guildId, enabled, minAccountAge, antiVpn, antiAlt, antiBot]
    VerificationConfigInput:
      type: object
      properties:
        enabled:
          type: boolean
        roleId:
          type: ["string", "null"]
        minAccountAge:
          type: integer
        antiVpn:
          type: boolean
        antiAlt:
          type: boolean
        antiBot:
          type: boolean
    VerificationResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        roleAssigned:
          type: boolean
      required: [success, message, roleAssigned]
...